Modified   include/DSP/Test.h Modified   include/DSP/Wah.h Modified   include/Wave/HighPassFilter.h Modified   include/Wave/LowPassFilter.h Modified   include/Wave/NoiseGate.h Modified   include/Wave/ToneControl.h Modified   include/Wave/WaveShaper.h

+++ include/Wave/NoiseGate.h
@@ -55,7 +55,7 @@ public:
 
 private:
 	BiquadFilter m_BiquadFilter;
-	float m_LowerThreshold;
-	float m_UpperThreshold;
+	double m_LowerThreshold;
+	double m_UpperThreshold;
 };
 #endif
